Move globe functions to separate file
