+++ src/geo/projection/globe_transform.ts
@@ -1,10 +1,9 @@
 import {mat2, mat4, vec3, vec4} from 'gl-matrix';
-import {MAX_VALID_LATITUDE, scaleZoom, TransformHelper, TransformUpdateResult} from '../transform_helper';
+import {MAX_VALID_LATITUDE, TransformHelper, TransformUpdateResult} from '../transform_helper';
 import {Tile} from '../../source/tile';
 import {MercatorTransform} from './mercator_transform';
 import {LngLat, earthRadius} from '../lng_lat';
-import {EXTENT} from '../../data/extent';
-import {clamp, differenceOfAnglesDegrees, distanceOfAnglesRadians, easeCubicInOut, lerp, mod, warnOnce, wrap} from '../../util/util';
+import {clamp, differenceOfAnglesDegrees, distanceOfAnglesRadians, easeCubicInOut, lerp, warnOnce} from '../../util/util';
 import {UnwrappedTileID, OverscaledTileID, CanonicalTileID} from '../../source/tile_id';
 import Point from '@mapbox/point-geometry';
 import {browser} from '../../util/browser';
@@ -17,20 +16,17 @@ import {LngLatBounds} from '../lng_lat_bounds';
 import {ITransform} from '../transform_interface';
 import {PaddingOptions} from '../edge_insets';
 import {translatePosition} from './mercator_utils';
+import {angularCoordinatesRadiansToVector, angularCoordinatesToSurfaceVector, getGlobeRadiusPixels, getZoomAdjustment, mercatorCoordinatesToAngularCoordinatesRadians, sphereSurfacePointToCoordinates} from './globe_utils';
+import {EXTENT} from '../../data/extent';
 
-export function getGlobeCircumferencePixels(transform: {worldSize: number; center: {lat: number}}): number {
-    const radius = getGlobeRadiusPixels(transform.worldSize, transform.center.lat);
-    const circumference = 2.0 * Math.PI * radius;
-    return circumference;
-}
-
-export function globeDistanceOfLocationsPixels(transform: {worldSize: number; center: {lat: number}}, a: LngLat, b: LngLat): number {
-    const vecA = angularCoordinatesToSurfaceVector(a);
-    const vecB = angularCoordinatesToSurfaceVector(b);
-    const dot = vec3.dot(vecA, vecB);
-    const radians = Math.acos(dot);
-    const circumference = getGlobeCircumferencePixels(transform);
-    return radians / (2.0 * Math.PI) * circumference;
+// These functions create **64** bit float vectors and matrices, unlike default gl-matrix functions.
+function createVec4(): vec4 { return new Float64Array(4) as any; }
+function createVec3(): vec3 { return new Float64Array(3) as any; }
+function createMat4(): mat4 { return new Float64Array(16) as any; }
+function createIdentityMat4(): mat4 {
+    const m = new Float64Array(16) as any;
+    mat4.identity(m);
+    return m;
 }
 
 /**
@@ -42,77 +38,6 @@ function tileCoordinatesToMercatorCoordinates(inTileX: number, inTileY: number, 
         inTileX / EXTENT * scale + tileID.canonical.x * scale,
         inTileY / EXTENT * scale + tileID.canonical.y * scale
     ];
-}
-
-/**
- * For given mercator coordinates in range 0..1, returns the angular coordinates on the sphere's surface, in radians.
- */
-export function mercatorCoordinatesToAngularCoordinatesRadians(mercatorX: number, mercatorY: number): [number, number] {
-    const sphericalX = mod(mercatorX * Math.PI * 2.0 + Math.PI, Math.PI * 2);
-    const sphericalY = 2.0 * Math.atan(Math.exp(Math.PI - (mercatorY * Math.PI * 2.0))) - Math.PI * 0.5;
-    return [sphericalX, sphericalY];
-}
-
-/**
- * For a given longitude and latitude (note: in radians) returns the normalized vector from the planet center to the specified place on the surface.
- * @param lngRadians - Longitude in radians.
- * @param latRadians - Latitude in radians.
- */
-export function angularCoordinatesRadiansToVector(lngRadians: number, latRadians: number): vec3 {
-    const len = Math.cos(latRadians);
-    const vec = createVec3();
-    vec[0] = Math.sin(lngRadians) * len;
-    vec[1] = Math.sin(latRadians);
-    vec[2] = Math.cos(lngRadians) * len;
-    return vec;
-}
-
-/**
- * For a given longitude and latitude (note: in degrees) returns the normalized vector from the planet center to the specified place on the surface.
- */
-export function angularCoordinatesToSurfaceVector(lngLat: LngLat): vec3 {
-    return angularCoordinatesRadiansToVector(lngLat.lng * Math.PI / 180, lngLat.lat * Math.PI / 180);
-}
-
-export function getGlobeRadiusPixels(worldSize: number, latitudeDegrees: number) {
-    // We want zoom levels to be consistent between globe and flat views.
-    // This means that the pixel size of features at the map center point
-    // should be the same for both globe and flat view.
-    // For this reason we scale the globe up when map center is nearer to the poles.
-    return worldSize / (2.0 * Math.PI) / Math.cos(latitudeDegrees * Math.PI / 180);
-}
-
-/**
- * Given a 3D point on the surface of a unit sphere, returns its angular coordinates in degrees.
- * The input vector must be normalized.
- */
-export function sphereSurfacePointToCoordinates(surface: vec3): LngLat {
-    const latRadians = Math.asin(surface[1]);
-    const latDegrees = latRadians / Math.PI * 180.0;
-    const lengthXZ = Math.sqrt(surface[0] * surface[0] + surface[2] * surface[2]);
-    if (lengthXZ > 1e-6) {
-        const projX = surface[0] / lengthXZ;
-        const projZ = surface[2] / lengthXZ;
-        const acosZ = Math.acos(projZ);
-        const lngRadians = (projX > 0) ? acosZ : -acosZ;
-        const lngDegrees = lngRadians / Math.PI * 180.0;
-        return new LngLat(wrap(lngDegrees, -180, 180), latDegrees);
-    } else {
-        return new LngLat(0.0, latDegrees);
-    }
-}
-
-/**
- * Computes how much to modify zoom to keep the globe size constant when changing latitude.
- * @param transform - An instance of any transform. Does not have any relation on the computed values.
- * @param oldLat - Latitude before change.
- * @param newLat - Latitude after change.
- * @returns A value to add to zoom level used for old latitude to keep same planet radius at new latitude.
- */
-export function getZoomAdjustment(oldLat: number, newLat: number): number {
-    const oldCircumference = Math.cos(oldLat * Math.PI / 180.0);
-    const newCircumference = Math.cos(newLat * Math.PI / 180.0);
-    return scaleZoom(newCircumference / oldCircumference);
 }
 
 /**
@@ -139,15 +64,6 @@ function angleToRotateBetweenVectors2D(vec1x: number, vec1y: number, vec2x: numb
     } else {
         return -angle;
     }
-}
-
-function createVec4(): vec4 { return new Float64Array(4) as any; }
-function createVec3(): vec3 { return new Float64Array(3) as any; }
-function createMat4(): mat4 { return new Float64Array(16) as any; }
-function createIdentityMat4(): mat4 {
-    const m = new Float64Array(16) as any;
-    mat4.identity(m);
-    return m;
 }
 
 export class GlobeTransform implements ITransform {
